EA-653 added in memory db for testing
