+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = thr.settings
+DJANGO_SETTINGS_MODULE = thr.settings.test
 python_files = tests.py test_*.py *_tests.py
 filterwarnings =
     ignore::DeprecationWarning
